09/03/2019 Debugging normal operation of multiple directory monitoring

+++ core/ServiceCore.hpp
@@ -246,7 +246,7 @@ int32_t ServiceCore::createEntryFile(const std::string &fileName, uint32_t value
 {
     std::string storageFile = appendBasePath(TRA_SYNC_FILE_NAME);
     std::string origFile = mLocalPath + fileName;
-    std::cout << "LHB my origFilePath = " << origFile << "value = 0x" << std::hex << value << std::endl;
+    // std::cout << "LHB my origFilePath = " << origFile << "value = 0x" << std::hex << value << std::endl;
     std::ofstream ouStream;
     int32_t rc = NO_ERROR;
     TRANSFER_ENTRY_FILE_T entry;
@@ -272,7 +272,7 @@ int32_t ServiceCore::createEntryFile(const std::string &fileName, uint32_t value
         switch (value) {
             case MONITOR_Removed:
                 entry.fileSize = 0;
-                std::cout << "LHB " << fileName << " value = " << value << std::endl;
+                // std::cout << "LHB " << fileName << " value = " << value << std::endl;
                 ouStream.write((char *)&entry, sizeof(TRANSFER_ENTRY_FILE_T));
                 break;
             case MONITOR_Updated:
@@ -286,7 +286,7 @@ int32_t ServiceCore::createEntryFile(const std::string &fileName, uint32_t value
                     }
                 }
                 if (SUCCEED(rc)) {
-                    std::cout << "LHB " << fileName << " value = " << value << std::endl;
+                    // std::cout << "LHB " << fileName << " value = " << value << std::endl;
                     entry.fileSize = inStream.tellg();
                     inStream.seekg(0);
                 }
@@ -347,18 +347,31 @@ int32_t ServiceCore::praseEntryFile(const std::string& inPath)
         while (!inStream.eof()) {
             memset(&entry, 0, sizeof(entry));
             inStream.read((char *)&entry, sizeof(TRANSFER_ENTRY_FILE_T));
-            LOGD(mModule, "LHB Flages = 0x%x 0x%x", entry.flages, MOENTRY_FLAGE_MASK);
+            //  LOGD(mModule, "LHB Flages = 0x%x 0x%x", entry.flages, MOENTRY_FLAGE_MASK);
             if (entry.flages != MOENTRY_FLAGE_MASK) {
                 std::cout << "Flage not match error\n";
                 break;
             }
 
+            /* accordance with filsPath first value to choose remote path */
+            std::string fileName = entry.fileName;
+            auto pos = fileName.find('/');
+            if (pos == fileName.npos) {
+                LOGE(mModule, "Can't find '/' in %s", fileName.c_str());
+                break;
+            }
+            auto key = fileName.substr(0, pos + 1);
+            auto relativeFilePath = fileName.substr(pos + 1);
+            // LOGI(mModule, "LHB Key=%s relativeFilePath=%s", key.c_str(), relativeFilePath.c_str());
+            auto absoluteFilePath = mRemoteDirNames[key] + relativeFilePath;
+            LOGI(mModule, "absoluteFilePath = %s", absoluteFilePath.c_str());
+
             switch (entry.value) {
                 case MONITOR_Removed: {
                     LOGD(mModule, "MONITOR_Removed runing...");
                     std::string cmdStr = "rm -rf ";
-                    cmdStr += mLocalPath;
-                    cmdStr +=  entry.fileName;
+                    // cmdStr += mLocalPath;
+                    cmdStr +=  absoluteFilePath;
                     system(cmdStr.c_str());
                 }
                 break;
@@ -375,13 +388,11 @@ int32_t ServiceCore::praseEntryFile(const std::string& inPath)
                     }
 
                     if (SUCCEED(rc)) {
-                        std::string tmpFilePath = mLocalPath;
-                        tmpFilePath += entry.fileName;
-                        LOGD(mModule, "SYNC file path = %s", tmpFilePath.c_str());
-                        std::ofstream ouStream(tmpFilePath, std::ios::binary | std::ios::trunc);
-                        \
+                        LOGD(mModule, "SYNC file path = %s", absoluteFilePath.c_str());
+                        std::ofstream ouStream(absoluteFilePath, std::ios::binary | std::ios::trunc);
+
                         if (!ouStream.is_open()) {
-                            LOGE(mModule, "Open %s failed\n", tmpFilePath.c_str());
+                            LOGE(mModule, "Open %s failed\n", absoluteFilePath.c_str());
                             rc = NOT_FOUND;
                         }
                         if (SUCCEED(rc)) {
